+++ lib/streamaggr/increase.go
@@ -2,7 +2,6 @@ package streamaggr
 
 import (
 	"fmt"
-	"math"
 
 	"github.com/VictoriaMetrics/metrics"
 
@@ -10,24 +9,23 @@ import (
 	"github.com/VictoriaMetrics/VictoriaMetrics/lib/fasttime"
 )
 
-type totalLastValue struct {
+type increaseLastValue struct {
 	value          float64
 	timestamp      int64
 	deleteDeadline int64
 }
 
-type totalAggrValueShared struct {
-	lastValues map[string]totalLastValue
-	total      float64
+type increaseAggrValueShared struct {
+	lastValues map[string]increaseLastValue
 }
 
-type totalAggrValue struct {
+type increaseAggrValue struct {
 	total  float64
-	shared *totalAggrValueShared
+	shared *increaseAggrValueShared
 }
 
-func (av *totalAggrValue) pushSample(c aggrConfig, sample *pushSample, key string, deleteDeadline int64) {
-	ac := c.(*totalAggrConfig)
+func (av *increaseAggrValue) pushSample(c aggrConfig, sample *pushSample, key string, deleteDeadline int64) {
+	ac := c.(*increaseAggrConfig)
 	currentTime := fasttime.UnixTimestamp()
 	keepFirstSample := ac.keepFirstSample && currentTime >= ac.ignoreFirstSampleDeadline
 	lv, ok := av.shared.lastValues[key]
@@ -53,10 +51,10 @@ func (av *totalAggrValue) pushSample(c aggrConfig, sample *pushSample, key strin
 	av.shared.lastValues[key] = lv
 }
 
-func (av *totalAggrValue) flush(c aggrConfig, ctx *flushCtx, key string, isLast bool) {
-	ac := c.(*totalAggrConfig)
+func (av *increaseAggrValue) flush(c aggrConfig, ctx *flushCtx, key string, isLast bool) {
+	ac := c.(*increaseAggrConfig)
 	suffix := ac.getSuffix()
-	total := av.shared.total + av.total
+	total := av.total
 	av.total = 0
 	lvs := av.shared.lastValues
 	for lk, lv := range lvs {
@@ -64,22 +62,16 @@ func (av *totalAggrValue) flush(c aggrConfig, ctx *flushCtx, key string, isLast 
 			delete(lvs, lk)
 		}
 	}
-	if math.Abs(total) >= (1 << 53) {
-		// It is time to reset the entry, since it starts losing float64 precision
-		av.shared.total = 0
-	} else {
-		av.shared.total = total
-	}
 	ctx.appendSeries(key, suffix, total)
 }
 
-func (av *totalAggrValue) state() any {
+func (av *increaseAggrValue) state() any {
 	return av.shared
 }
 
-func newTotalAggrConfig(ms *metrics.Set, metricLabels string, ignoreFirstSampleIntervalSecs uint64, keepFirstSample bool) aggrConfig {
+func newIncreaseAggrConfig(ms *metrics.Set, metricLabels string, ignoreFirstSampleIntervalSecs uint64, keepFirstSample bool) aggrConfig {
 	ignoreFirstSampleDeadline := fasttime.UnixTimestamp() + ignoreFirstSampleIntervalSecs
-	cfg := &totalAggrConfig{
+	cfg := &increaseAggrConfig{
 		keepFirstSample:           keepFirstSample,
 		ignoreFirstSampleDeadline: ignoreFirstSampleDeadline,
 	}
@@ -87,34 +79,31 @@ func newTotalAggrConfig(ms *metrics.Set, metricLabels string, ignoreFirstSampleI
 	return cfg
 }
 
-type totalAggrConfig struct {
-	// Whether to take into account the first sample in new time series when calculating the output value.
+type increaseAggrConfig struct {
 	keepFirstSample bool
 
 	// The first sample per each new series is ignored until this unix timestamp deadline in seconds even if keepFirstSample is set.
-	// This allows avoiding an initial spike of the output values at startup when new time series
-	// cannot be distinguished from already existing series. This is tracked with ignoreFirstSampleDeadline.
 	ignoreFirstSampleDeadline uint64
 	counterResetsTotal        *metrics.Counter
 }
 
-func (*totalAggrConfig) getValue(s any) aggrValue {
-	var shared *totalAggrValueShared
+func (*increaseAggrConfig) getValue(s any) aggrValue {
+	var shared *increaseAggrValueShared
 	if s == nil {
-		shared = &totalAggrValueShared{
-			lastValues: make(map[string]totalLastValue),
+		shared = &increaseAggrValueShared{
+			lastValues: make(map[string]increaseLastValue),
 		}
 	} else {
-		shared = s.(*totalAggrValueShared)
+		shared = s.(*increaseAggrValueShared)
 	}
-	return &totalAggrValue{
+	return &increaseAggrValue{
 		shared: shared,
 	}
 }
 
-func (ac *totalAggrConfig) getSuffix() string {
+func (ac *increaseAggrConfig) getSuffix() string {
 	if ac.keepFirstSample {
-		return "total"
+		return "increase"
 	}
-	return "total_prometheus"
+	return "increase_prometheus"
 }